# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HyperClaw API",
    "version": "1.2.4",
    "description": "HyperClaw REST API baseline. See docs/api.md for details."
  },
  "servers": [
    { "url": "http://127.0.0.1:8790", "description": "Local development" }
  ],
  "paths": {
...
    "/api/tasks/{id}/inject": {
      "post": {
        "summary": "Queue interrupt injection prompt for paused task",
        "responses": {
          "200": { "description": "Injection queued" },
          "400": { "description": "Invalid status or empty prompt" },
          "403": { "description": "Invalid interrupt token" }
        }
      }
    },
...
  }
}